dashboard/backend/requirements.txt
flask==3.0.0
flask-socketio==5.3.6
eventlet==0.33.3
paho-mqtt==1.6.1
PyYAML==6.0.1
opencv-python==4.9.0.80
mediapipe==0.10.14
numpy==1.26.4
